Added Character appearence and disappearence + Fixed many things
Fixed the revealing platform, now you can set the platforms to be revealed and it will make them hidden in its begin play, also allows other actors to if needed. the timer works and is in the game mode. Added character appearance and disappearance and applied an effect to a duplicate material. added a death and respawn feature to the bp_thirdperson character for testing (press R to have a look at the material in work) :)